+++ src/sds200/themes/home-assistant/sds200-cards.js
@@ -1,3 +1,3 @@
 import "/local/sds200/sds200-card.js?v=beb1c6f22d62655caf4fc541a0cabfa4ed273b8fe22d6b3fe4324f5dc88ab9d8";
 import "/local/sds200/sds200-display-card.js?v=b2d47c2b7abd19a92b2ee61b6b3de00362366f8df828d7786c54ae35aa0ada72";
-import "/local/sds200/sds200-waterfall-card.js?v=a9913f9d29528a489dcbd0370ca2d2ba656481b68fc0f48712719879a1214dcc";
+import "/local/sds200/sds200-waterfall-card.js?v=d850fa81b04b1798dc7e7f947737525d3a58538f106202f66384eb4e028e62d8";